requirements.txt
asgiref==3.7.2
Brotli==1.1.0
dj-database-url==2.1.0
Django==5.0
django-ckeditor==6.7.0
django-js-asset==2.2.0
gunicorn==21.2.0
packaging==23.2
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.6.0
